+++ Modules/Intent.Modules.EntityFrameworkCore.DesignTimeDbContextFactory/Templates/DesignTimeDbContextFactory/DesignTimeDbContextFactoryTemplatePartial.cs
@@ -47,17 +47,17 @@ public partial class DesignTimeDbContextFactoryTemplate : CSharpTemplateBase<obj
                 @class.AddMethod(GetDbContextName(), "CreateDbContext", method =>
                 {
                     method.AddParameter("string[]", "args");
-                    var primaryConnectionStringName = ExecutionContext.Settings.GetDatabaseSettings().PrimaryConnectionStringName();
-                    if (string.IsNullOrWhiteSpace(primaryConnectionStringName))
+                    var defaultConnectionStringName = ExecutionContext.Settings.GetDatabaseSettings().DefaultConnectionStringName();
+                    if (string.IsNullOrWhiteSpace(defaultConnectionStringName))
                     {
-                        primaryConnectionStringName = "DefaultConnection";
+                        defaultConnectionStringName = "DefaultConnection";
                     }
 
                     method.XmlComments.AddStatements($@"
 /// <inheritdoc />
 /// <param name=""args"">
 /// This is optional but will only accept 1 parameter which is the name of the connection string to lookup
-/// in a local appsettings.json file. By default this will use ""{primaryConnectionStringName}"".
+/// in a local appsettings.json file. By default this will use ""{defaultConnectionStringName}"".
 /// </param>");
                     if (ExecutionContext.Settings.GetDatabaseSettings().DatabaseProvider().IsInMemory())
                     {
@@ -74,7 +74,7 @@ public partial class DesignTimeDbContextFactoryTemplate : CSharpTemplateBase<obj
                         .AddChainStatement("Build()"));
                     method.AddStatement("var connStringName = args.FirstOrDefault();");
                     method.AddIfStatement("string.IsNullOrEmpty(connStringName)", stmt => stmt
-                        .AddStatement($@"connStringName = ""{primaryConnectionStringName}"";"));
+                        .AddStatement($@"connStringName = ""{defaultConnectionStringName}"";"));
 
                     const string connectionStringStatement = "var connectionString = configuration.GetConnectionString(connStringName);";